refactor(global): Change global project structure in order to split concerns + some style improvement
