+++ buildConfig/config.js
