+++ buildConfig/config.remoteaws.js
