+++ src/containers/page-home/page-home.spec.js
@@ -1,9 +1,9 @@
-jest.dontMock('./index.jsx');
+jest.dontMock('./page-questionnaire.jsx');
 
 import React from 'react';
 import { shallow } from 'enzyme';
 
-import { PageHome } from './index';
+import { PageHome } from './page-home';
 import { getLocale } from 'utils/test/test-utils';
 
 const locale = getLocale();
