+++ src/index.ejs
@@ -3,7 +3,7 @@
   <head>
     <meta charset="utf-8">
     <meta name="viewport" content="user-scalable=no,initial-scale=1,maximum-scale=1,minimum-scale=1,width=device-width">
-    <link rel="icon" href="img/favicon.ico"/>
+    <link rel="icon" href="scss/images/favicon.ico"/>
     <title>Pogues</title>
   </head>
   <body>
